+++ sonic-framework/wrappers/vst3wf/source/dsp/SynthesizerBase.h
@@ -20,16 +20,16 @@ protected:
 
 public:
   virtual ~ParameterBuilder() = default;
-  virtual void addUnary(uint64_t address, Str identifier, Str label,
+  virtual void addUnary(uint32_t address, Str identifier, Str label,
                         double defaultValue, Str group = "",
                         ParameterFlags flags = ParameterFlags::None) = 0;
-  virtual void addEnum(uint64_t address, Str identifier, Str label,
+  virtual void addEnum(uint32_t address, Str identifier, Str label,
                        Str defaultValueString, StrVec valueStrings,
                        Str group = "",
                        ParameterFlags flags = ParameterFlags::None) = 0;
-  // virtual void addEnum(uint64_t address, Str identifier, Str label,
+  // virtual void addEnum(uint32_t address, Str identifier, Str label,
   //                      int defaultKey, std::map<int, Str> valueOptions) = 0;
-  virtual void addBool(uint64_t address, Str identifier, Str label,
+  virtual void addBool(uint32_t address, Str identifier, Str label,
                        bool defaultValue, Str group = "",
                        ParameterFlags flags = ParameterFlags::None) = 0;
 };
@@ -38,7 +38,7 @@ class SynthesizerBase {
 public:
   virtual ~SynthesizerBase() = default;
   virtual void setupParameters(ParameterBuilder &builder) = 0;
-  virtual void setParameter(uint64_t address, double value) = 0;
+  virtual void setParameter(uint32_t address, double value) = 0;
   virtual void prepare(double sampleRate, int32_t maxFrameCount) = 0;
   virtual void noteOn(int32_t noteNumber, double velocity) = 0;
   virtual void noteOff(int32_t noteNumber) = 0;
